FIX : Scene Implementation

+++ src/main/java/net/biryeongtrain/text_emulator/level/scene/SceneSelector.java
@@ -8,14 +8,14 @@ public record SceneSelector(Condition condition,
                             Operator operator,
                             String value,
                             Unit unit,
-                            float weight) {
+                            int weight) {
 
     public static Codec<SceneSelector> CODEC = RecordCodecBuilder.create(instance -> instance.group(
             Condition.CODEC.fieldOf("condition").forGetter(SceneSelector::condition),
             Operator.CODEC.fieldOf("operator").forGetter(SceneSelector::operator),
             Codec.STRING.fieldOf("value").forGetter(SceneSelector::value),
             Unit.CODEC.fieldOf("unit").forGetter(SceneSelector::unit),
-            Codec.FLOAT.fieldOf("weight").forGetter(SceneSelector::weight)
+            Codec.INT.fieldOf("weight").forGetter(SceneSelector::weight)
     ).apply(instance, SceneSelector::new));
